Messaging related cleanup
