+++ src/languageTools/LanguageClientWrapper.js
@@ -26,8 +26,160 @@
 define(function (require, exports, module) {
     "use strict";
 
-    var MessageHandler = require("languageTools/MessageHandler"),
-        ToolingInfo = JSON.parse(require("text!languageTools/ToolingInfo.json"));
+    var ToolingInfo = JSON.parse(require("text!languageTools/ToolingInfo.json")),
+		MESSAGE_FORMAT = {
+			BRACKETS : "brackets",
+			LSP : "lsp"
+		};
+
+    function _addTypeInformation(type, params) {
+        return {
+            type: type,
+            params: params
+        };
+    }
+
+    function hasValidProp(obj, prop) {
+        return (obj && obj[prop] !== undefined && obj[prop] !== null);
+    }
+
+    function hasValidProps(obj, props) {
+        var retval = !!obj,
+            len = props.length,
+            i;
+
+        for (i = 0; retval && (i < len); i ++) {
+            retval = (retval && obj[props[i]] !== undefined && obj[props[i]] !== null);
+        }
+
+        return retval;
+    }
+    /*
+        RequestParams creator - sendNotifications/request
+    */
+    function validateRequestParams(type, params) {
+        var validatedParams = null;
+
+        params = params ? params : {};
+
+        //Don't validate if the formatting is done by the caller
+        if (params.format === MESSAGE_FORMAT.LSP) {
+            return params;
+        }
+
+        switch (type) {
+        case ToolingInfo.LANGUAGE_SERVICE.START:
+            {
+                if (hasValidProp(params, "rootPath")) {
+                    validatedParams = params;
+                    validatedParams.capabilities = validatedParams.capabilities || false;
+                }
+                break;
+            }
+        case ToolingInfo.FEATURES.CODE_HINTS:
+        case ToolingInfo.FEATURES.PARAMETER_HINTS:
+        case ToolingInfo.FEATURES.JUMP_TO_DECLARATION:
+        case ToolingInfo.FEATURES.JUMP_TO_DEFINITION:
+        case ToolingInfo.FEATURES.JUMP_TO_IMPL:
+            {
+                if (hasValidProps(params, ["filePath", "cursorPos"])) {
+                    validatedParams = params;
+                }
+                break;
+            }
+        case ToolingInfo.FEATURES.CODE_HINT_INFO:
+            {
+                validatedParams = params;
+                break;
+            }
+        case ToolingInfo.FEATURES.FIND_REFERENCES:
+            {
+                if (hasValidProps(params, ["filePath", "cursorPos"])) {
+                    validatedParams = params;
+                    validatedParams.includeDeclaration = validatedParams.includeDeclaration || false;
+                }
+                break;
+            }
+        case ToolingInfo.FEATURES.DOCUMENT_SYMBOLS:
+            {
+                if (hasValidProp(params, "filePath")) {
+                    validatedParams = params;
+                }
+                break;
+            }
+        case ToolingInfo.FEATURES.PROJECT_SYMBOLS:
+            {
+                if (params && params.query && typeof params.query === "string") {
+                    validatedParams = params;
+                }
+                break;
+            }
+        case ToolingInfo.LANGUAGE_SERVICE.CUSTOM_REQUEST:
+            {
+                validatedParams = params;
+            }
+        }
+
+        return validatedParams;
+    }
+
+    /*
+        ReponseParams transformer - used by OnNotifications
+    */
+    function validateNotificationParams(type, params) {
+        var validatedParams = null;
+
+        params = params ? params : {};
+
+        //Don't validate if the formatting is done by the caller
+        if (params.format === "spec") {
+            return params;
+        }
+
+        switch (type) {
+        case ToolingInfo.SYNCHRONIZE_EVENTS.DOCUMENT_OPENED:
+            {
+                if (hasValidProps(params, ["filePath", "fileContent", "languageId"])) {
+                    validatedParams = params;
+                }
+                break;
+            }
+        case ToolingInfo.SYNCHRONIZE_EVENTS.DOCUMENT_CHANGED:
+            {
+                if (hasValidProps(params, ["filePath", "fileContent"])) {
+                    validatedParams = params;
+                }
+                break;
+            }
+        case ToolingInfo.SYNCHRONIZE_EVENTS.DOCUMENT_SAVED:
+            {
+                if (hasValidProp(params, "filePath")) {
+                    validatedParams = params;
+                }
+                break;
+            }
+        case ToolingInfo.SYNCHRONIZE_EVENTS.DOCUMENT_CLOSED:
+            {
+                if (hasValidProp(params, "filePath")) {
+                    validatedParams = params;
+                }
+                break;
+            }
+        case ToolingInfo.SYNCHRONIZE_EVENTS.PROJECT_FOLDERS_CHANGED:
+            {
+                if (hasValidProps(params, ["foldersAdded", "foldersRemoved"])) {
+                    validatedParams = params;
+                }
+                break;
+            }
+        case ToolingInfo.LANGUAGE_SERVICE.CUSTOM_NOTIFICATION:
+            {
+                validatedParams = params;
+            }
+        }
+
+        return validatedParams;
+    }
 
     function validateHandler(handler) {
         var retval = false;
@@ -119,8 +271,9 @@ define(function (require, exports, module) {
     };
 
     LanguageClientWrapper.prototype._request = function (type, params) {
-        params = MessageHandler.constructRequestParams(type, params);
+        params = validateRequestParams(type, params);
         if (params) {
+            params = _addTypeInformation(type, params);
             return this._requestClient(params);
         }
 
@@ -129,8 +282,9 @@ define(function (require, exports, module) {
     };
 
     LanguageClientWrapper.prototype._notify = function (type, params) {
-        params = MessageHandler.constructNotificationParams(type, params);
+        params = validateNotificationParams(type, params);
         if (params) {
+            params = _addTypeInformation(type, params);
             this._notifyClient(params);
         } else {
             console.log("Invalid Parameters provided for notification type : ", type);
@@ -158,7 +312,7 @@ define(function (require, exports, module) {
     */
     //start
     LanguageClientWrapper.prototype.start = function (params) {
-        params = MessageHandler.constructRequestParams(ToolingInfo.LANGUAGE_SERVICE.START, params);
+        params = validateRequestParams(ToolingInfo.LANGUAGE_SERVICE.START, params);
         return this._startClient(params);
     };
