+++ src/languageTools/MessageHandler.js
@@ -46,155 +46,6 @@ define(function (require, exports, module) {
         return new NodeDomain(domainName, domainPath);
     }
 
-    function _createParams(type, params) {
-        return {
-            type: type,
-            params: params
-        };
-    }
-
-    function hasValidProp(obj, prop) {
-        return (obj && obj[prop] !== undefined && obj[prop] !== null);
-    }
-
-    function hasValidProps(obj, props) {
-        var retval = !!obj,
-            len = props.length,
-            i;
-
-        for (i = 0; retval && (i < len); i ++) {
-            retval = (retval && obj[props[i]] !== undefined && obj[props[i]] !== null);
-        }
-
-        return retval;
-    }
-    /*
-        RequestParams creator - sendNotifications/request
-    */
-    function constructRequestParams(type, params) {
-        var jsonParams = null;
-
-        params = params ? params : {};
-
-        //Don't construct if the formatting is done by the caller
-        if (params.format === "spec") {
-            return params;
-        }
-
-        switch (type) {
-        case ToolingInfo.LANGUAGE_SERVICE.START:
-            {
-                if (hasValidProp(params, "rootPath")) {
-                    jsonParams = _createParams(type, params);
-                    jsonParams["capabilities"] = params.capabilities ? params.capabilities : false;
-                }
-                break;
-            }
-        case ToolingInfo.FEATURES.CODE_HINTS:
-        case ToolingInfo.FEATURES.PARAMETER_HINTS:
-        case ToolingInfo.FEATURES.JUMP_TO_DECLARATION:
-        case ToolingInfo.FEATURES.JUMP_TO_DEFINITION:
-        case ToolingInfo.FEATURES.JUMP_TO_IMPL:
-            {
-                if (hasValidProps(params, ["filePath", "cursorPos"])) {
-                    jsonParams = _createParams(type, params);
-                }
-                break;
-            }
-        case ToolingInfo.FEATURES.CODE_HINT_INFO:
-            {
-                jsonParams = _createParams(type, params);
-                break;
-            }
-        case ToolingInfo.FEATURES.FIND_REFERENCES:
-            {
-                if (hasValidProps(params, ["filePath", "cursorPos"])) {
-                    jsonParams = _createParams(type, params);
-                    jsonParams["includeDeclaration"] = params.includeDeclaration ? params.includeDeclaration : false;
-                }
-                break;
-            }
-        case ToolingInfo.FEATURES.DOCUMENT_SYMBOLS:
-            {
-                if (hasValidProp(params, "filePath")) {
-                    jsonParams = _createParams(type, params);
-                }
-                break;
-            }
-        case ToolingInfo.FEATURES.PROJECT_SYMBOLS:
-            {
-                if (params && params.query && typeof params.query === "string") {
-                    jsonParams = _createParams(type, params);
-                }
-                break;
-            }
-        case ToolingInfo.LANGUAGE_SERVICE.CUSTOM_REQUEST:
-            {
-                jsonParams = _createParams(type, params);
-            }
-        }
-
-        return jsonParams;
-    }
-
-    /*
-        ReponseParams transformer - used by OnNotifications
-    */
-    function constructNotificationParams(type, params) {
-        var jsonParams = null;
-
-        params = params ? params : {};
-
-        //Don't construct if the formatting is done by the caller
-        if (params.format === "spec") {
-            return params;
-        }
-
-        switch (type) {
-        case ToolingInfo.SYNCHRONIZE_EVENTS.DOCUMENT_OPENED:
-            {
-                if (hasValidProps(params, ["filePath", "fileContent", "languageId"])) {
-                    jsonParams = _createParams(type, params);
-                }
-                break;
-            }
-        case ToolingInfo.SYNCHRONIZE_EVENTS.DOCUMENT_CHANGED:
-            {
-                if (hasValidProps(params, ["filePath", "fileContent"])) {
-                    jsonParams = _createParams(type, params);
-                }
-                break;
-            }
-        case ToolingInfo.SYNCHRONIZE_EVENTS.DOCUMENT_SAVED:
-            {
-                if (hasValidProp(params, "filePath")) {
-                    jsonParams = _createParams(type, params);
-                }
-                break;
-            }
-        case ToolingInfo.SYNCHRONIZE_EVENTS.DOCUMENT_CLOSED:
-            {
-                if (hasValidProp(params, "filePath")) {
-                    jsonParams = _createParams(type, params);
-                }
-                break;
-            }
-        case ToolingInfo.SYNCHRONIZE_EVENTS.PROJECT_FOLDERS_CHANGED:
-            {
-                if (hasValidProps(params, ["foldersAdded", "foldersRemoved"])) {
-                    jsonParams = _createParams(type, params);
-                }
-                break;
-            }
-        case ToolingInfo.LANGUAGE_SERVICE.CUSTOM_NOTIFICATION:
-            {
-                jsonParams = _createParams(type, params);
-            }
-        }
-
-        return jsonParams;
-    }
-
     function loadLanguageClientDomain(clientName, domainPath) {
         //generate a random hash name for the domain, this is the client id
         var domainName = clientName,
@@ -242,6 +93,4 @@ define(function (require, exports, module) {
     }
 
     exports.initiateLanguageClient = initiateLanguageClient;
-    exports.constructRequestParams = constructRequestParams;
-    exports.constructNotificationParams = constructNotificationParams;
 });
